+++ askbot/conf/settings_wrapper.py
@@ -174,12 +174,7 @@ class ConfigSettings(object):
 
     def as_dict(self):
         cache_key = get_bulk_cache_key()
-        settings = cache.get(cache_key)
-        if settings:
-            return settings
-        else:
-            self.prime_cache(cache_key)
-            return cache.get(cache_key)
+        return cache.get(cache_key) or self.prime_cache(cache_key)
 
     @classmethod
     def prime_cache(cls, cache_key, **kwargs):
@@ -194,6 +189,7 @@ class ConfigSettings(object):
             else:
                 out[key] = hardcoded_setting
         cache.set(cache_key, out)
+        return out
 
 
 def get_bulk_cache_key():