+++ askbot/media/js/post.js
@@ -138,9 +138,11 @@ var CPValidator = (function () {
                     limit_tag_length: true
                 },
                 text: {
+                    required: !!askbot.settings.minQuestionBodyLength,
                     minlength: askbot.settings.minQuestionBodyLength
                 },
                 title: {
+                    required: true,
                     minlength: askbot.settings.minTitleLength
                 }
             };
@@ -184,6 +186,7 @@ var CPValidator = (function () {
         getAnswerFormRules: function () {
             return {
                 text: {
+                    required: true,
                     minlength: askbot.settings.minAnswerBodyLength
                 }
             };
@@ -315,6 +318,10 @@ MergeQuestionsDialog.prototype.setPreview = function (data) {
     this._preview.fadeIn();
 };
 
+MergeQuestionsDialog.prototype.clearIdInput = function () {
+    this._idInput.val('');
+};
+
 MergeQuestionsDialog.prototype.clearPreview = function () {
     for (var i = 0; i < this._tags.length; i++) {
         this._tags[i].dispose();
@@ -322,6 +329,7 @@ MergeQuestionsDialog.prototype.clearPreview = function () {
     this._previewTitle.html('');
     this._previewBody.html('');
     this._previewTags.html('');
+    this.setAcceptButtonText(gettext('Load preview'));
     this._preview.hide();
 };
 
@@ -349,7 +357,7 @@ MergeQuestionsDialog.prototype.getLoadPreviewHandler = function () {
         // here I am disabling eqeqeq because it looks like there's a type coercion going on, can't be sure
         // so skipping it
         /*jshint eqeqeq:false*/
-        if (curId && curId != prevId) {
+        if (curId) {// && curId != prevId) {
         /*jshint eqeqeq:true*/
             $.ajax({
                 type: 'GET',
@@ -360,13 +368,48 @@ MergeQuestionsDialog.prototype.getLoadPreviewHandler = function () {
                     me.setPreview(data);
                     me.setPrevToId(curId);
                     me.setAcceptButtonText(gettext('Merge'));
+                    me.setPreviewLoaded(true);
+                    return false;
                 },
                 error: function () {
                     me.clearPreview();
                     me.setAcceptButtonText(gettext('Load preview'));
+                    me.setPreviewLoaded(false);
+                    return false;
                 }
             });
         }
+    };
+};
+
+MergeQuestionsDialog.prototype.setPreviewLoaded = function(isLoaded) {
+    this._isPreviewLoaded = isLoaded;
+};
+
+MergeQuestionsDialog.prototype.isPreviewLoaded = function() {
+    return this._isPreviewLoaded;
+};
+
+MergeQuestionsDialog.prototype.getAcceptHandler = function() {
+    var me = this;
+    return function() {
+        if (me.isPreviewLoaded()) {
+            var handler = me.getStartMergingHandler();
+        } else {
+            var handler = me.getLoadPreviewHandler();
+        }
+        handler();
+        return false;
+    };
+};
+
+MergeQuestionsDialog.prototype.getRejectHandler = function() {
+    var me = this;
+    return function() {
+        me.clearPreview();
+        me.clearIdInput();
+        me.setPreviewLoaded(false);
+        me.hide();
     };
 };
 
@@ -417,12 +460,13 @@ MergeQuestionsDialog.prototype.createDom = function () {
     this.setRejectButtonText(gettext('Cancel'));
     this.setAcceptButtonText(gettext('Load preview'));
     this.setHeadingText(askbot.messages.mergeQuestions);
-    this.setAcceptHandler(this.getStartMergingHandler());
+    this.setRejectHandler(this.getRejectHandler());
+    this.setAcceptHandler(this.getAcceptHandler());
 
     MergeQuestionsDialog.superClass_.createDom.call(this);
     this._element.hide();
 
-    this._fromId = $('.post.question').data('postId');
+    this._fromId = $('.js-question').data('postId');
     //have to do this on document since _element is not in the DOM yet
     $(document).trigger('askbot.afterMergeQuestionsDialogCreateDom', [this]);
 };