+++ askbot/models/__init__.py
@@ -28,7 +28,7 @@ from django.db import models
 from django.db.models import Count
 from django.conf import settings as django_settings
 from django.contrib.contenttypes.models import ContentType
-from django.core import cache
+from django.core.cache import cache
 from django.core import exceptions as django_exceptions
 from django_countries.fields import CountryField
 from askbot import exceptions as askbot_exceptions
@@ -59,6 +59,7 @@ from askbot.models.meta import ImportRun, ImportedObjectInfo
 from askbot import auth
 from askbot.utils.decorators import auto_now_timestamp
 from askbot.utils.decorators import reject_forbidden_phrases
+from askbot.utils.cache import memoize, delete_memoized
 from askbot.utils.markup import URL_RE
 from askbot.utils.slug import slugify
 from askbot.utils.html import replace_links_with_text
@@ -188,11 +189,18 @@ User.add_to_class('reputation',
 )
 User.add_to_class('gravatar', models.CharField(max_length=32))
 #User.add_to_class('has_custom_avatar', models.BooleanField(default=False))
+
 User.add_to_class(
     'avatar_type',
-    models.CharField(max_length=1,
-        choices=const.AVATAR_STATUS_CHOICE,
-        default='n')
+    models.CharField(
+        max_length=1,
+        choices=(
+            ('n', _('None')),
+            ('g', _('Gravatar')),#only if user has real uploaded gravatar
+            ('a', _('Uploaded Avatar')),#avatar uploaded locally - with django-avatar app
+        ),
+        default='n'
+    )
 )
 User.add_to_class('gold', models.SmallIntegerField(default=0))
 User.add_to_class('silver', models.SmallIntegerField(default=0))
@@ -282,34 +290,60 @@ def user_get_default_avatar_url(self, size):
     """
     return skin_utils.get_media_url(askbot_settings.DEFAULT_AVATAR_URL)
 
+def user_get_avatar_type(self):
+    """returns user avatar type, taking into account
+    avatar_type value and how use of avatar and/or gravatar
+    is configured
+    Value returned is one of 'n', 'a', 'g'.
+    """
+    if 'avatar' in django_settings.INSTALLED_APPS:
+        if self.avatar_type == 'g':
+            if askbot_settings.ENABLE_GRAVATAR:
+                return 'g'
+            else:
+                #fallback to default avatar if gravatar is disabled
+                return 'n'
+        assert(self.avatar_type in ('a', 'n'))#only these are allowed
+        return self.avatar_type
+
+    #if we don't have an uploaded avatar, always use gravatar
+    return 'g'
+
+@memoize
 def user_get_avatar_url(self, size=48):
     """returns avatar url - by default - gravatar,
     but if application django-avatar is installed
     it will use avatar provided through that app
     """
-    if 'avatar' in django_settings.INSTALLED_APPS:
-        if self.avatar_type == 'n':
-            import avatar
-            if askbot_settings.ENABLE_GRAVATAR: #avatar.settings.AVATAR_GRAVATAR_BACKUP:
-                return self.get_gravatar_url(size)
-            else:
-                return self.get_default_avatar_url(size)
-        elif self.avatar_type == 'a':
-            kwargs = {'user_id': self.id, 'size': size}
-            try:
-                return reverse('avatar_render_primary', kwargs = kwargs)
-            except NoReverseMatch:
-                message = 'Please, make sure that avatar urls are in the urls.py '\
-                          'or update your django-avatar app, '\
-                          'currently it is impossible to serve avatars.'
-                logging.critical(message)
-                raise django_exceptions.ImproperlyConfigured(message)
-        else:
-            return self.get_gravatar_url(size)
-    if askbot_settings.ENABLE_GRAVATAR:
-        return self.get_gravatar_url(size)
-    else:
+    avatar_type = self.get_avatar_type()
+
+    if avatar_type == 'n':
         return self.get_default_avatar_url(size)
+    elif avatar_type == 'a':
+        from avatar.conf import settings as avatar_settings
+        sizes = avatar_settings.AUTO_GENERATE_AVATAR_SIZES
+        if size not in sizes:
+            logging.critical('add values %d to setting AUTO_GENERATE_AVATAR_SIZES')
+
+        kwargs = {'user': self.username, 'size': size}
+        try:
+            return reverse('avatar_render_primary', kwargs=kwargs)
+        except NoReverseMatch:
+            message = 'Please, make sure that avatar urls are in the urls.py '\
+                      'or update your django-avatar app, '\
+                      'currently it is impossible to serve avatars.'
+            logging.critical(message)
+            raise django_exceptions.ImproperlyConfigured(message)
+    assert(avatar_type == 'g')
+    return self.get_gravatar_url(size)
+
+
+def user_clear_avatar_cache(self):
+    from avatar.conf import settings as avatar_settings
+    sizes = avatar_settings.AUTO_GENERATE_AVATAR_SIZES
+    for size in sizes:
+        delete_memoized(user_get_avatar_url, self, size=size)
+
 
 def user_get_top_answers_paginator(self, visitor=None):
     """get paginator for top answers by the user for a
@@ -1311,8 +1345,10 @@ def user_assert_can_revoke_old_vote(self, vote):
     """raises exceptions.PermissionDenied if old vote
     cannot be revoked due to age of the vote
     """
-    if (datetime.datetime.now().day - vote.voted_at.day) \
-        >= askbot_settings.MAX_DAYS_TO_CANCEL_VOTE:
+    if askbot_settings.MAX_DAYS_TO_CANCEL_VOTE < 0:
+        return
+    if (datetime.datetime.now() - vote.voted_at).days \
+            >= askbot_settings.MAX_DAYS_TO_CANCEL_VOTE:
         raise django_exceptions.PermissionDenied(
             _('sorry, but older votes cannot be revoked')
         )
@@ -3247,7 +3283,9 @@ User.add_to_class(
     user_subscribe_for_followed_question_alerts
 )
 User.add_to_class('get_absolute_url', user_get_absolute_url)
+User.add_to_class('get_avatar_type', user_get_avatar_type)
 User.add_to_class('get_avatar_url', user_get_avatar_url)
+User.add_to_class('clear_avatar_cache', user_clear_avatar_cache)
 User.add_to_class('get_default_avatar_url', user_get_default_avatar_url)
 User.add_to_class('get_gravatar_url', user_get_gravatar_url)
 User.add_to_class('get_or_create_fake_user', user_get_or_create_fake_user)
@@ -3888,7 +3926,6 @@ def make_admin_if_first_user(user, **kwargs):
 
     function is run when user registers
     """
-    import sys
     user_count = User.objects.all().count()
     if user_count == 1:
         user.set_status('d')
@@ -3999,7 +4036,7 @@ def record_spam_rejection(
         activity.active_at = now
         activity.summary = summary
         activity.save()
-    
+
 
 from south.signals import ran_migration 
 
@@ -4101,6 +4138,7 @@ if 'avatar' in django_settings.INSTALLED_APPS:
         sender=Avatar,
         dispatch_uid='update_avatar_type_flag_on_avatar_delete'
     )
+
 
 django_signals.post_delete.connect(
     record_cancel_vote,